+++ hivemind/client/averaging/key_manager.py
